Update redirect handling and add async support
Modified the `RedirectUri` in the `Challenge` method of `HomeController` to use a general root URL with a token query parameter for improved flexibility. Added a `using` directive for `System.Threading.Tasks` in `Program.cs` to support asynchronous programming. Implemented an event handler for `OnRedirectToIdentityProvider` in the `Program` class to enhance the authentication flow by controlling the redirection process during OpenID Connect authentication.